app/pyproject.toml
[tool.poetry]
name = "take-home"
version = "0.1.0"
description = "Basic Django web application for take home assignment."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.3"
psycopg2-binary = "^2.9.6"
requests = "^2.31.0"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]

[[tool.poetry.source]]
name = "fpho"
url = "https://files.pythonhosted.org"
priority = "primary"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
